Separated codes to files by function or page
