+++ lib/screens/gallery_page.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+
+class GalleryPage extends StatelessWidget {
+  final List<String> galleryItems = List.generate(
+    30,
+    (index) => 'ギャラリーアイテム ${index + 1}',
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('アルバム'),
+        centerTitle: true,
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8.0),
+        child: GridView.builder(
+          gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+            crossAxisCount: 2,
+            crossAxisSpacing: 10,
+            mainAxisSpacing: 10,
+            childAspectRatio: 1.0,
+          ),
+          itemCount: galleryItems.length,
+          itemBuilder: (context, index) {
+            return Card(
+              elevation: 0,
+              shape: RoundedRectangleBorder(
+                side: BorderSide(color: Colors.grey.shade300),
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Padding(
+                padding: const EdgeInsets.all(12.0),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Flexible(
+                      child: Text(
+                        galleryItems[index],
+                        style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    ),
+                    SizedBox(height: 8),
+                    Expanded(
+                      child: Text(
+                        'これはギャラリーアイテム ${index + 1} の説明文です。ここに詳細情報が入ります。',
+                        style: TextStyle(fontSize: 14),
+                        overflow: TextOverflow.fade,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
