+++ lib/screens/setting_detail_page.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+
+class SettingDetailPage extends StatelessWidget {
+  final String title;
+  final String description;
+
+  const SettingDetailPage({
+    required this.title,
+    required this.description,
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(title),
+        centerTitle: true,
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              title,
+              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+            ),
+            SizedBox(height: 16),
+            Text(
+              description,
+              style: TextStyle(fontSize: 16),
+            ),
+            SizedBox(height: 24),
+            Divider(),
+            SizedBox(height: 16),
+            Text(
+              'このセクションの詳細設定はここに表示されます。現在開発中です。',
+              style: TextStyle(fontSize: 14, color: Colors.grey[600]),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
