+++ pubspec.yaml
@@ -1,7 +1,7 @@
-name: namer_app
-description: A new Flutter project.
+name: picture_reminder
+description: An app reminding you to take a picture of your day.
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
 version: 0.0.1+1
